minor fixes

+++ lib/headers/utilities.h
@@ -8,7 +8,7 @@
 #include <ostream>
 
 float ran1(long *idum);
-inline int ran1(const int &min, const int &max,long *idum)
-{return (max - min)*ran1(idum) + min;}
+inline float ran1(const int &min, const int &max,long *idum)
+{return static_cast<float>((max - min)*ran1(idum) + min);}
 
 #endif /* __UTILITIES_H */
